Move star styling into scss, camel case table
