+++ apps/public/package.json
@@ -36,6 +36,7 @@
     "@types/react": "^16.9.19",
     "concurrently": "^5.0.2",
     "cypress": "^3.8.2",
+    "js-levenshtein": "^1.1.6",
     "next-transpile-modules": "^3.0.1",
     "sass-loader": "^8.0.2",
     "typescript": "^3.7.4",
