+++ shared/ui-components/src/page_components/UnitTables.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import { storiesOf } from "@storybook/react"
 import UnitTables from "./UnitTables"
-import { BasicTable } from "../tables/basic_table"
+import { BasicTable } from "../tables/BasicTable"
 import { GroupedTable } from "../tables/GroupedTable"
 import Archer from "@bloom-housing/listings-service/listings/archer.json"
 import { unitSummariesTable, groupNonReservedAndReservedSummaries } from "../helpers/tableSummaries"
