+++ shared/ui-components/src/page_components/UnitTables.tsx
@@ -2,7 +2,7 @@ import * as React from "react"
 import nanoid from "nanoid"
 import { MinMax } from "@bloom-housing/core/src/general"
 import { UnitSummary, Unit } from "@bloom-housing/core/src/units"
-import { BasicTable } from "../../src/tables/basic_table"
+import { BasicTable } from "../tables/BasicTable"
 import t from "../../src/helpers/translator"
 import numberOrdinal from "../../src/helpers/numberOrdinal"
 import capitalize from "../../src/helpers/capitalize"
